requirements.txt
python-dotenv==1.0.0
pynetbox==7.2.0
colorama==0.4.6
paramiko==3.3.1
prettytable==3.9.0
netaddr==0.9.0